requirement.txt
fastapi
uvicorn
python-multipart
pillow
numpy==1.26.4
torch
torchvision
torchaudio
opencv-python
mediapipe==0.10.20
scikit-learn
protobuf==4.25.4
scikit-image
